# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api batista",
        "description": "A api fornece e guarda dados relacionados as igrejas, bem como dados de membros, pastores e etc.. ",
        "contact": {
            "name": "Danilo",
            "email": "daniloganda95@gmail.com"
        },
        "license": {
            "name": "batista",
            "url": "#"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "faz o login",
                "operationId": "postUserLogin",
                "requestBody": {
                    "description": "User object precisa para fazer o login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "format": "email",
                                        "default": "daniloganda95@gmail.com"
                                    },
                                    "password": {
                                        "description": "senha do usuário",
                                        "type": "string",
                                        "default": "Danilo123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "parametro inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Unauthorized",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "model Member",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Authorized",
                                            "type": "string"
                                        },
                                        "token": {
                                            "description": "token de acesso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "faz o logout",
                "operationId": "getUserLogout",
                "responses": {
                    "200": {
                        "description": "model Member",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "sucesso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}